Add sleep as workaround to decrease CPU at tbl0901
Add time sleeps as workaround to decrease CPU load at tbl0901.

Apparently the XMController is performing around a thousand of
queries per second to tbl0901 Tango Database. This increases
the CPU of tbl0901 machine to around 60-70% and the load to 3-4.
The load should be ideally less than 1.

By using sleeps of 100ms, as the MotorInterface devices lasts 100ms
to answer, the next query is not done, till the answer is there.
This allows to decrease the CPU load.

This is a workaround. The real problem it seems to be
located to the XMController SW. It could be because many
DeviceProxies are created in TangoInterface.dll of XMController,
each time that a Tango Command or that a Tango Attribute is read.

However a mystery remains: for each creation of a DeviceProxy,
we see that 2 entries in the admin blackbox of the tbl0901 device
(seen in jive). If we perform a 'mytop' command as root in the
machine tbl0901, we see that the number of queries for each
DeviceProxy creation is between 4 and 8. Really far away from
the 1000 queries per second that we saw with spusso if doing a
'mytop' in tbl0901 machine (we should be seeing
4motors*8queries/deviceproxy = 32 queries per second at most).